[Chore/#72] 답변 확인 뷰 뷰모델 분리 및 데이터 모델 분리
